requirements.txt
langdetect==1.0.9
langcodes==3.4.0
langtrans==1.0.6
PyYAML==6.0.2
langchain==0.2.16
langchain-ollama==0.1.3
fastapi==0.111.0
uvicorn==0.30.1
openai==1.40.0
